finally added some javadoc
Parts of the javadoc is inspired from the maven-surefire-plugin since
it's the same kind of plugin. The javadoc for RunMojo is mandatory in
order to produce a comprehensive help goal.

Again, some refactoring. With a little more work, classes implementing
commons-exec interfaces could be moved to an independant artefact
without any coupling to Maven.