added a minimal configuration integration test
The test builds a war that is deployed in a tomcat server during the
integration-test phases. The maven-invoker-plugin configuration creates
a self-contained local repository for the integration tests and makes
the builds offline to ensure reproductibility. There is a parent pom for
the it modules to make common configuration (eg. plugins versions)
reusable.