varnishtest run mojo learned how to write reports
It has also been decided that Java 6 at least would be required since
it's already reached its EOL. By the way, the mojo finally learned how
to handle timeouts... The mojo also found a new home with a more obvious
package name.

On the pure execution side, it happens that the parsing of the test
title was plain wrong, this is fixed.