fixed the broken maven site generation, *sigh*
It's actually not possible to work with only artifacts from the reactor.
The maven-site-plugin (or one of its friends, I don't know) will search
for the artifacts currently being built in the local repository. It
doesn't matter that they're being built and available in the target/
directory, and that the artifacts built from the reactor are likely to
be more recent. It has to go through the local repository, period.